[refactor] #258 어댑터와 이넘 클래스의 패키지 위치 변경
