+++ app/src/main/java/org/go/sopt/winey/util/binding/BindingAdapter.kt
@@ -19,7 +19,7 @@ import coil.load
 import coil.transform.RoundedCornersTransformation
 import de.hdodenhof.circleimageview.CircleImageView
 import org.go.sopt.winey.R
-import org.go.sopt.winey.presentation.main.feed.WineyFeedType
+import org.go.sopt.winey.presentation.model.WineyFeedType
 import org.go.sopt.winey.presentation.model.UserLevel
 import org.go.sopt.winey.presentation.nickname.NicknameActivity.Companion.MY_PAGE_SCREEN
 import org.go.sopt.winey.presentation.nickname.NicknameActivity.Companion.STORY_SCREEN
